mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.2.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
]
